requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
commonmark==0.9.1
discord.py==1.7.3
gunicorn==20.1.0
idna==3.3
multidict==6.0.2
Pygments==2.12.0
python-decouple==3.6
requests==2.28.1
rich==12.5.1
typing_extensions==4.3.0
urllib3==1.26.10
yarl==1.7.2

# - disocrd.py (stable)
# - gunicorn for heroku 
# - requests for http requests
# - python decouple for config files
# - rich for color text in terminal

